+++ src/style/components/generics/_index.scss
@@ -12,4 +12,3 @@
 @use 'playlist_entry_minimal';
 @use 'tab_bar';
 @use 'token_widget';
-@use 'user_widget';
