+++ src/style/components/user/_index.scss
@@ -5,3 +5,4 @@
 //
 
 @use 'user';
+@use 'widgets';
