+++ src/style/components/user/widgets/_index.scss
@@ -1,0 +1,7 @@
+//
+// Dakara Project
+//
+// User widgets main page style file
+//
+
+@use 'user';
