+++ src/style/components/user/widgets/_user.scss
